// ui/models/generated/token.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A token of the system
 */
export interface Token {
  /**
   * The ID of the token
   */
  id: string;
  /**
   * The name of the token
   */
  name: string;
  /**
   * The description of the token
   */
  description?: string;
  /**
   * The display name of the token
   */
  displayName?: string;
  /**
   * The date and time the token was updated
   */
  updatedAt?: string;
  /**
   * The date and time the token was created
   */
  createdAt?: string;
  /**
   * The token
   */
  token: string;
  /**
   * The type of the token
   */
  tokenType: string;
  /**
   * The date and time the token will expire
   */
  expirationDate: string;
  /**
   * The ID of the user
   */
  userId: string;
  /**
   * The ID of the device
   */
  deviceId?: string;
  /**
   * The IP address of the device
   */
  ip?: string;
}